# Use getAncestors in getDescendant