Both `getAncestors` and `getDescendant` are DFS, but the `getDescendant` cannot take advantage by the cache by `getAncestors`.   This PR makes `getDescendant` taking advantage from `getAncestors` cache.