@ianstormtaylor 2018-02-05
@zhujinxuan can you run the benchmarks and take a screenshot of the results? Also could you edit the comment to have a `PERF:` prefix and explain why we're doing this?