@zhujinxuan 2018-02-05
@ianstormtaylor   Currently, there is perhaps a bug in benchmark:
https://github.com/ianstormtaylor/slate/issues/1606