@zhujinxuan 2018-02-05
```
benchmarks
    html-serializer
      deserialize
        8.47 → 7.80 ops/sec
      serialize
        1114.36 → 883.52 ops/sec
    plain-serializer
      deserialize
        450.37 → 536.30 ops/sec
      serialize
        8646.61 → 9727.72 ops/sec
    rendering
      normal
        103.56 → 108.65 ops/sec
    changes
      delete-backward
        110856.88 → 95104.75 ops/sec
      delete-forward
        19438.07 → 15309.81 ops/sec
      insert-text-by-key-multiple
        90.14 → 84.36 ops/sec
      insert-text-by-key
        986.45 → 1103.69 ops/sec
      insert-text
        1157.27 → 1182.72 ops/sec
      normalize
        1820.24 → 2150.49 ops/sec
      split-block
        26.24 → 33.07 ops/sec
    models
      from-json
        61.86 → 139.17 ops/sec (125% faster) 🙌
      get-blocks-at-range
        617076.02 → 741345.62 ops/sec
      get-blocks
        2694134.71 → 2799980.24 ops/sec
      get-characters-at-range
        824954.03 → 892295.84 ops/sec
      get-characters
        3037965.33 → 3062649.37 ops/sec
      get-inlines-at-range
        782431.04 → 749365.01 ops/sec
      get-inlines
        3326124.17 → 3086753.69 ops/sec
      get-leaves
        2123041.39 → 1999890.00 ops/sec
      get-marks-at-range
        812624.58 → 700687.20 ops/sec
      get-marks
        3308037.37 → 3056577.60 ops/sec
      get-path
        944887.06 → 882762.53 ops/sec
      get-texts-at-range
        894863.43 → 915673.12 ops/sec
      get-texts
        3061266.17 → 3293666.18 ops/sec
      has-node-multiple
        108403.03 → 127487.55 ops/sec
      has-node
        836775.36 → 938308.89 ops/sec
      to-json
        3136.12 → 3417.45 ops/sec
      update-node
        18503.33 → 19358.83 ops/sec
```
Benchmark test after fixing benchmark in https://github.com/ianstormtaylor/slate/pull/1607