@zhujinxuan 2018-02-05
Benchmark result:  
But I cannot understand why from-json becomes faster.
```
benchmarks
    html-serializer
      deserialize
        8.47 → 9.19 ops/sec
      serialize
        1114.36 → 865.70 ops/sec
    plain-serializer
      deserialize
        450.37 → 409.61 ops/sec
      serialize
        8646.61 → 9775.98 ops/sec
    rendering
      normal
        103.56 → 115.02 ops/sec
    changes
      delete-backward
        110856.88 → 110168.10 ops/sec
      delete-forward
        19438.07 → 26458.53 ops/sec (36% faster)
      insert-text-by-key-multiple
        90.14 → 91.01 ops/sec
      insert-text-by-key
        986.45 → 1252.71 ops/sec
      insert-text
        1157.27 → 1171.50 ops/sec
      normalize
        1820.24 → 1803.57 ops/sec
      split-block
        26.24 → 20.75 ops/sec
    models
      from-json
        61.86 → 155.29 ops/sec (151% faster) 🙌
      get-blocks-at-range
        617076.02 → 786911.54 ops/sec
      get-blocks
        2694134.71 → 2918882.87 ops/sec
      get-characters-at-range
        824954.03 → 832815.08 ops/sec
      get-characters
        3037965.33 → 3048774.43 ops/sec
      get-inlines-at-range
        782431.04 → 788504.04 ops/sec
      get-inlines
        3326124.17 → 3193938.12 ops/sec
      get-leaves
        2123041.39 → 2112337.97 ops/sec
      get-marks-at-range
        812624.58 → 915376.99 ops/sec
      get-marks
        3308037.37 → 3023311.08 ops/sec
      get-path
        944887.06 → 951787.17 ops/sec
      get-texts-at-range
        894863.43 → 910620.71 ops/sec
      get-texts
        3061266.17 → 3085942.85 ops/sec
      has-node-multiple
        108403.03 → 119431.55 ops/sec
      has-node
        836775.36 → 1016023.01 ops/sec
      to-json
        3136.12 → 3310.83 ops/sec
      update-node
        18503.33 → 15678.18 ops/sec

✨  Done in 66.59s.
```