@zhujinxuan 2018-02-05
Second Run, quite a different result of benchmark:
```
  benchmarks
    html-serializer
      deserialize
        8.67 → 9.38 ops/sec
      serialize
        1165.78 → 1157.79 ops/sec
    plain-serializer
      deserialize
        463.89 → 628.33 ops/sec (35% faster)
      serialize
        7566.29 → 8903.07 ops/sec
    rendering
      normal
        118.04 → 115.29 ops/sec
    changes
      delete-backward
        111887.21 → 107278.07 ops/sec
      delete-forward
        18023.80 → 29838.32 ops/sec (66% faster)
      insert-text-by-key-multiple
        94.65 → 123.25 ops/sec
      insert-text-by-key
        1063.50 → 1130.89 ops/sec
      insert-text
        1047.14 → 1229.03 ops/sec
      normalize
        2096.03 → 1706.62 ops/sec
      split-block
        29.40 → 31.81 ops/sec
    models
      from-json
        90.28 → 83.09 ops/sec
      get-blocks-at-range
        912672.92 → 616782.87 ops/sec
      get-blocks
        3533857.20 → 3250202.32 ops/sec
      get-characters-at-range
        951137.66 → 778152.22 ops/sec
      get-characters
        3239059.67 → 2666184.34 ops/sec
      get-inlines-at-range
        890236.70 → 749539.01 ops/sec
      get-inlines
        3242335.62 → 3096977.41 ops/sec
      get-leaves
        2288355.28 → 1895259.41 ops/sec
      get-marks-at-range
        896903.10 → 821223.98 ops/sec
      get-marks
        3352056.82 → 3313233.90 ops/sec
      get-path
        1004350.81 → 741941.02 ops/sec
      get-texts-at-range
        914519.12 → 752961.80 ops/sec
      get-texts
        3116750.99 → 3341709.34 ops/sec
      has-node-multiple
        128690.85 → 117821.53 ops/sec
      has-node
        901851.32 → 915632.21 ops/sec
      to-json
        3266.99 → 3503.94 ops/sec
      update-node
        19941.43 → 16198.22 ops/sec
```